+++ src/TurtlePath.EntityFrameworkCore/TurtlePathDbContextOptions.cs
@@ -1,6 +1,7 @@
 namespace TurtlePath.EntityFrameworkCore
 {
     using System.Reflection;
+    using TurtlePath.Domain.Identifier;
 
     /// <summary>
     /// Configures the TurtlePath conventions applied by <see cref="BaseDbContext"/>.
@@ -26,6 +27,11 @@ namespace TurtlePath.EntityFrameworkCore
         /// Gets a value indicating whether CId value converters should be applied.
         /// </summary>
         public bool ApplyCIdConverters { get; init; } = true;
+
+        /// <summary>
+        /// Gets the identifier definition used to configure CId database conversion.
+        /// </summary>
+        public CIdDefinition CIdDefinition { get; init; }
 
         /// <summary>
         /// Gets the assemblies used to discover entity configurations. When empty, the DbContext assembly is used.