+++ src/TurtlePath/TurtlePathServiceCollectionExtensions.cs
@@ -1,6 +1,7 @@
 namespace Microsoft.Extensions.DependencyInjection
 {
     using System.Reflection;
+    using TurtlePath;
     using TurtlePath.Domain.Identifier;
     using TurtlePath.Hooks;
 
@@ -15,7 +16,7 @@ namespace Microsoft.Extensions.DependencyInjection
         /// <param name="services">The service collection.</param>
         /// <param name="hookAssemblies">Assemblies that contain TurtlePath hooks.</param>
         /// <returns>The same service collection.</returns>
-        public static IServiceCollection AddTurtlePath(this IServiceCollection services, params Assembly[] hookAssemblies)
+        public static ITurtlePathBuilder AddTurtlePath(this IServiceCollection services, params Assembly[] hookAssemblies)
         {
             if (services == null)
                 throw new ArgumentNullException(nameof(services));
@@ -23,7 +24,7 @@ namespace Microsoft.Extensions.DependencyInjection
             if (hookAssemblies?.Length > 0)
                 services.AddHandlerHooksFromAssemblies(hookAssemblies);
 
-            return services;
+            return new TurtlePathBuilder(services);
         }
 
         /// <summary>
@@ -35,7 +36,7 @@ namespace Microsoft.Extensions.DependencyInjection
         /// <param name="configureIdentifier">The CId configuration callback.</param>
         /// <param name="hookAssemblies">Assemblies that contain TurtlePath hooks.</param>
         /// <returns>The same service collection.</returns>
-        public static IServiceCollection AddTurtlePath<TTargetType, TDbType>(
+        public static ITurtlePathBuilder AddTurtlePath<TTargetType, TDbType>(
             this IServiceCollection services,
             Action<CIdConfiguration<TTargetType, TDbType>> configureIdentifier,
             params Assembly[] hookAssemblies)
@@ -43,10 +44,30 @@ namespace Microsoft.Extensions.DependencyInjection
             if (services == null)
                 throw new ArgumentNullException(nameof(services));
 
-            services.UseCId(configureIdentifier);
-            services.AddTurtlePath(hookAssemblies);
+            var builder = services.AddTurtlePath(hookAssemblies);
+            builder.UseCId(configureIdentifier);
 
-            return services;
+            return builder;
+        }
+
+        /// <summary>
+        /// Configures TurtlePath identifiers on the current registration pipeline.
+        /// </summary>
+        /// <typeparam name="TTargetType">The domain identifier value type.</typeparam>
+        /// <typeparam name="TDbType">The database identifier value type.</typeparam>
+        /// <param name="builder">The TurtlePath builder.</param>
+        /// <param name="configureIdentifier">The CId configuration callback.</param>
+        /// <returns>The same TurtlePath builder.</returns>
+        public static ITurtlePathBuilder UseCId<TTargetType, TDbType>(
+            this ITurtlePathBuilder builder,
+            Action<CIdConfiguration<TTargetType, TDbType>> configureIdentifier)
+        {
+            if (builder == null)
+                throw new ArgumentNullException(nameof(builder));
+
+            builder.Services.UseCId(configureIdentifier);
+
+            return builder;
         }
     }
 }